+++ saleor/graphql/checkout/mutations/checkout_create.py
@@ -11,6 +11,7 @@ from ....product import models as product_models
 from ....warehouse.reservations import get_reservation_length, is_reservation_enabled
 from ...account.i18n import I18nMixin
 from ...account.types import AddressInput
+from ...app.dataloaders import load_app
 from ...channel.utils import clean_channel
 from ...core.descriptions import (
     ADDED_IN_31,
@@ -168,7 +169,8 @@ class CheckoutCreate(ModelMutation, I18nMixin):
     def clean_checkout_lines(
         cls, info, lines, country, channel
     ) -> Tuple[List[product_models.ProductVariant], List["CheckoutLineData"]]:
-        check_permissions_for_custom_prices(info.context.app, lines)
+        app = load_app(info.context)
+        check_permissions_for_custom_prices(app, lines)
         variant_ids = [line["variant_id"] for line in lines]
         variants = cls.get_nodes_or_error(
             variant_ids,