+++ saleor/graphql/checkout/mutations/checkout_lines_update.py
@@ -5,6 +5,7 @@ from django.forms import ValidationError
 
 from ....checkout.error_codes import CheckoutErrorCode
 from ....warehouse.reservations import is_reservation_enabled
+from ...app.dataloaders import load_app
 from ...checkout.types import CheckoutLine
 from ...core.descriptions import (
     ADDED_IN_31,
@@ -129,8 +130,9 @@ class CheckoutLinesUpdate(CheckoutLinesAdd):
         discounts,
         replace,
     ):
+        app = load_app(info.context)
         # if the requestor is not app, the quantity is required for all lines
-        if not info.context.app:
+        if not app:
             if any(
                 [
                     line_data.quantity_to_update is False