+++ saleor/graphql/invoice/mutations.py
@@ -7,6 +7,7 @@ from ...invoice import events, models
 from ...invoice.error_codes import InvoiceErrorCode
 from ...invoice.notifications import send_invoice
 from ...order import events as order_events
+from ..app.dataloaders import load_app
 from ..core.mutations import ModelDeleteMutation, ModelMutation
 from ..core.types import InvoiceError
 from ..order.types import Order
@@ -83,22 +84,22 @@ class InvoiceRequest(ModelMutation):
         invoice = info.context.plugins.invoice_request(
             order=order, invoice=shallow_invoice, number=data.get("number")
         )
-
+        app = load_app(info.context)
         if invoice and invoice.status == JobStatus.SUCCESS:
             order_events.invoice_generated_event(
                 order=order,
                 user=info.context.user,
-                app=info.context.app,
+                app=app,
                 invoice_number=invoice.number,
             )
         else:
             order_events.invoice_requested_event(
-                user=info.context.user, app=info.context.app, order=order
+                user=info.context.user, app=app, order=order
             )
 
         events.invoice_requested_event(
             user=info.context.user,
-            app=info.context.app,
+            app=app,
             order=order,
             number=data.get("number"),
         )
@@ -173,9 +174,10 @@ class InvoiceCreate(ModelMutation):
         invoice.order = order
         invoice.status = JobStatus.SUCCESS
         invoice.save()
+        app = load_app(info.context)
         events.invoice_created_event(
             user=info.context.user,
-            app=info.context.app,
+            app=app,
             invoice=invoice,
             number=cleaned_input["number"],
             url=cleaned_input["url"],
@@ -183,7 +185,7 @@ class InvoiceCreate(ModelMutation):
         order_events.invoice_generated_event(
             order=order,
             user=info.context.user,
-            app=info.context.app,
+            app=app,
             invoice_number=cleaned_input["number"],
         )
         return InvoiceCreate(invoice=invoice)
@@ -209,8 +211,9 @@ class InvoiceRequestDelete(ModelMutation):
         invoice.status = JobStatus.PENDING
         invoice.save(update_fields=["status", "updated_at"])
         info.context.plugins.invoice_delete(invoice)
+        app = load_app(info.context)
         events.invoice_requested_deletion_event(
-            user=info.context.user, app=info.context.app, invoice=invoice
+            user=info.context.user, app=app, invoice=invoice
         )
         return InvoiceRequestDelete(invoice=invoice)
 
@@ -231,8 +234,9 @@ class InvoiceDelete(ModelDeleteMutation):
     def perform_mutation(cls, _root, info, **data):
         invoice = cls.get_instance(info, **data)
         response = super().perform_mutation(_root, info, **data)
+        app = load_app(info.context)
         events.invoice_deleted_event(
-            user=info.context.user, app=info.context.app, invoice_id=invoice.pk
+            user=info.context.user, app=app, invoice_id=invoice.pk
         )
         return response
 
@@ -288,10 +292,11 @@ class InvoiceUpdate(ModelMutation):
         )
         instance.status = JobStatus.SUCCESS
         instance.save(update_fields=["external_url", "number", "updated_at", "status"])
+        app = load_app(info.context)
         order_events.invoice_updated_event(
             order=instance.order,
             user=info.context.user,
-            app=info.context.app,
+            app=app,
             invoice_number=instance.number,
             url=instance.url,
             status=instance.status,
@@ -342,7 +347,11 @@ class InvoiceSendNotification(ModelMutation):
     def perform_mutation(cls, _root, info, **data):
         instance = cls.get_instance(info, **data)
         cls.clean_instance(info, instance)
+        app = load_app(info.context)
         send_invoice(
-            instance, info.context.user, info.context.app, info.context.plugins
+            invoice=instance,
+            staff_user=info.context.user,
+            app=app,
+            manager=info.context.plugins,
         )
         return InvoiceSendNotification(invoice=instance)