+++ saleor/graphql/order/mutations/order_discount_delete.py
@@ -5,6 +5,7 @@ from ....core.tracing import traced_atomic_transaction
 from ....order import events
 from ....order.search import update_order_search_vector
 from ....order.utils import invalidate_order_prices, remove_order_discount_from_order
+from ...app.dataloaders import load_app
 from ...core.types import OrderError
 from ..types import Order
 from .order_discount_common import OrderDiscountCommon
@@ -31,13 +32,15 @@ class OrderDiscountDelete(OrderDiscountCommon):
             info, data.get("discount_id"), only_type="OrderDiscount"
         )
         order = order_discount.order
+        app = load_app(info.context)
+
         cls.validate_order(info, order)
 
         remove_order_discount_from_order(order, order_discount)
         events.order_discount_deleted_event(
             order=order,
             user=info.context.user,
-            app=info.context.app,
+            app=app,
             order_discount=order_discount,
         )