+++ saleor/graphql/order/mutations/order_refund.py
@@ -8,6 +8,7 @@ from ....order.actions import order_refunded
 from ....order.error_codes import OrderErrorCode
 from ....payment import PaymentError, TransactionKind, gateway
 from ....payment.gateway import request_refund_action
+from ...app.dataloaders import load_app
 from ...core.mutations import BaseMutation
 from ...core.scalars import PositiveDecimal
 from ...core.types import OrderError
@@ -69,6 +70,7 @@ class OrderRefund(BaseMutation):
 
         order = cls.get_node_or_error(info, data.get("id"), only_type=Order)
         clean_order_refund(order)
+        app = load_app(info.context)
 
         if payment_transactions := list(order.payment_transactions.all()):
             # We use the last transaction as we don't have a possibility to
@@ -80,7 +82,7 @@ class OrderRefund(BaseMutation):
                     refund_value=amount,
                     channel_slug=order.channel.slug,
                     user=info.context.user,
-                    app=info.context.app,
+                    app=app,
                 )
             except PaymentError as e:
                 raise ValidationError(
@@ -94,7 +96,7 @@ class OrderRefund(BaseMutation):
             transaction = try_payment_action(
                 order,
                 info.context.user,
-                info.context.app,
+                app,
                 payment,
                 gateway.refund,
                 payment,
@@ -108,7 +110,7 @@ class OrderRefund(BaseMutation):
                 order_refunded(
                     order,
                     info.context.user,
-                    info.context.app,
+                    app,
                     amount,
                     payment,
                     info.context.plugins,