+++ saleor/graphql/product/mutations/products.py
@@ -33,6 +33,7 @@ from ....product.utils import delete_categories, get_products_ids_without_varian
 from ....product.utils.variants import generate_and_set_variant_name
 from ....thumbnail import models as thumbnail_models
 from ....warehouse.management import deactivate_preorder_for_variant
+from ...app.dataloaders import load_app
 from ...attribute.types import AttributeValueInput
 from ...attribute.utils import AttributeAssignmentMixin, AttrValuesInput
 from ...channel import ChannelContext
@@ -764,10 +765,11 @@ class ProductDelete(ModelDeleteMutation):
             pk__in=draft_order_lines_data.line_pks
         ).delete()
 
+        app = load_app(info.context)
         # run order event for deleted lines
         for order, order_lines in draft_order_lines_data.order_to_lines_mapping.items():
             order_events.order_line_product_removed_event(
-                order, info.context.user, info.context.app, order_lines
+                order, info.context.user, app, order_lines
             )
 
         order_pks = draft_order_lines_data.order_pks
@@ -1180,9 +1182,10 @@ class ProductVariantDelete(ModelDeleteMutation):
         ).delete()
 
         # run order event for deleted lines
+        app = load_app(info.context)
         for order, order_lines in draft_order_lines_data.order_to_lines_mapping.items():
             order_events.order_line_variant_removed_event(
-                order, info.context.user, info.context.app, order_lines
+                order, info.context.user, app, order_lines
             )
 
         order_pks = draft_order_lines_data.order_pks