+++ saleor/order/tests/test_order.py
@@ -90,7 +90,13 @@ def test_recalculate_order_keeps_weight_unit(order_with_lines):
 
 
 def test_add_variant_to_order_adds_line_for_new_variant(
-    order_with_lines, product, product_translation_fr, settings, info, site_settings
+    order_with_lines,
+    product,
+    product_translation_fr,
+    settings,
+    anonymous_user,
+    anonymous_plugins,
+    site_settings,
 ):
     order = order_with_lines
     variant = product.variants.get()
@@ -99,12 +105,12 @@ def test_add_variant_to_order_adds_line_for_new_variant(
     line_data = OrderLineData(variant_id=str(variant.id), variant=variant, quantity=1)
 
     add_variant_to_order(
-        order,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
+        order=order,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
     )
 
     line = order.lines.last()
@@ -128,7 +134,8 @@ def test_add_variant_to_order_adds_line_for_new_variant_on_sale(
     sale,
     discount_info,
     settings,
-    info,
+    anonymous_user,
+    anonymous_plugins,
     site_settings,
 ):
     order = order_with_lines
@@ -140,13 +147,13 @@ def test_add_variant_to_order_adds_line_for_new_variant_on_sale(
     line_data = OrderLineData(variant_id=str(variant.id), variant=variant, quantity=1)
 
     add_variant_to_order(
-        order,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
-        [discount_info],
+        order=order,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
+        discounts=[discount_info],
     )
 
     line = order.lines.last()
@@ -171,7 +178,13 @@ def test_add_variant_to_order_adds_line_for_new_variant_on_sale(
 
 
 def test_add_variant_to_draft_order_adds_line_for_variant_with_price_0(
-    order_with_lines, product, product_translation_fr, settings, info, site_settings
+    order_with_lines,
+    product,
+    product_translation_fr,
+    settings,
+    anonymous_user,
+    anonymous_plugins,
+    site_settings,
 ):
     order = order_with_lines
     variant = product.variants.get()
@@ -184,12 +197,12 @@ def test_add_variant_to_draft_order_adds_line_for_variant_with_price_0(
     line_data = OrderLineData(variant_id=str(variant.id), variant=variant, quantity=1)
 
     add_variant_to_order(
-        order,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
+        order=order,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
     )
 
     line = order.lines.last()
@@ -203,7 +216,11 @@ def test_add_variant_to_draft_order_adds_line_for_variant_with_price_0(
 
 
 def test_add_variant_to_order_not_allocates_stock_for_new_variant(
-    order_with_lines, product, info, site_settings
+    order_with_lines,
+    product,
+    anonymous_user,
+    anonymous_plugins,
+    site_settings,
 ):
     variant = product.variants.get()
     stock = Stock.objects.get(product_variant=variant)
@@ -212,12 +229,12 @@ def test_add_variant_to_order_not_allocates_stock_for_new_variant(
 
     line_data = OrderLineData(variant_id=str(variant.id), variant=variant, quantity=1)
     add_variant_to_order(
-        order_with_lines,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
+        order=order_with_lines,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
     )
 
     stock.refresh_from_db()
@@ -225,7 +242,7 @@ def test_add_variant_to_order_not_allocates_stock_for_new_variant(
 
 
 def test_add_variant_to_order_edits_line_for_existing_variant(
-    order_with_lines, info, site_settings
+    order_with_lines, anonymous_user, anonymous_plugins, site_settings
 ):
     existing_line = order_with_lines.lines.first()
     variant = existing_line.variant
@@ -236,12 +253,12 @@ def test_add_variant_to_order_edits_line_for_existing_variant(
     )
 
     add_variant_to_order(
-        order_with_lines,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
+        order=order_with_lines,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
     )
 
     existing_line.refresh_from_db()
@@ -252,7 +269,7 @@ def test_add_variant_to_order_edits_line_for_existing_variant(
 
 
 def test_add_variant_to_order_not_allocates_stock_for_existing_variant(
-    order_with_lines, info, site_settings
+    order_with_lines, anonymous_user, anonymous_plugins, site_settings
 ):
     existing_line = order_with_lines.lines.first()
     variant = existing_line.variant
@@ -265,12 +282,12 @@ def test_add_variant_to_order_not_allocates_stock_for_existing_variant(
     )
 
     add_variant_to_order(
-        order_with_lines,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
+        order=order_with_lines,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
     )
 
     stock.refresh_from_db()
@@ -542,17 +559,19 @@ def test_calculate_order_weight(order_with_lines):
     assert calculated_weight == order_weight
 
 
-def test_order_weight_add_more_variant(order_with_lines, info, site_settings):
+def test_order_weight_add_more_variant(
+    order_with_lines, anonymous_user, anonymous_plugins, site_settings
+):
     variant = order_with_lines.lines.first().variant
     line_data = OrderLineData(variant_id=str(variant.id), variant=variant, quantity=2)
 
     add_variant_to_order(
-        order_with_lines,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
+        order=order_with_lines,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
     )
     order_with_lines.refresh_from_db()
 
@@ -561,17 +580,23 @@ def test_order_weight_add_more_variant(order_with_lines, info, site_settings):
     )
 
 
-def test_order_weight_add_new_variant(order_with_lines, product, info, site_settings):
+def test_order_weight_add_new_variant(
+    order_with_lines,
+    product,
+    anonymous_user,
+    anonymous_plugins,
+    site_settings,
+):
     variant = product.variants.first()
     line_data = OrderLineData(variant_id=str(variant.id), variant=variant, quantity=2)
 
     add_variant_to_order(
-        order_with_lines,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
+        order=order_with_lines,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
     )
     order_with_lines.refresh_from_db()
 
@@ -605,7 +630,11 @@ def test_order_weight_delete_line(lines_info):
 
 
 def test_get_order_weight_non_existing_product(
-    order_with_lines, product, info, site_settings
+    order_with_lines,
+    product,
+    anonymous_user,
+    anonymous_plugins,
+    site_settings,
 ):
     # Removing product should not affect order's weight
     order = order_with_lines
@@ -613,12 +642,12 @@ def test_get_order_weight_non_existing_product(
     line_data = OrderLineData(variant_id=str(variant.id), variant=variant, quantity=1)
 
     add_variant_to_order(
-        order,
-        line_data,
-        info.context.user,
-        info.context.app,
-        info.context.plugins,
-        site_settings,
+        order=order,
+        line_data=line_data,
+        user=anonymous_user,
+        app=None,
+        manager=anonymous_plugins,
+        site_settings=site_settings,
     )
     old_weight = order.get_total_weight()