+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable, unreleased changes to this project will be documented in this file.
 # 3.7.0 [Unreleased]
 
 ### Other changes
+
 - Fix situation when Payment Gateway try to save to long error message - #10402 by @fowczarek
+- Replaced `context.app` lazy object with a dataloader.
 - Add support for bcrypt password hashes - #10346 by @pkucmus
+
 
 # 3.6.0